requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/dev.in
#
-e git+https://github.com/liip-forks/django-agpl#egg=django-agpl
    # via -r /code/requirements/base.in
-e git+https://github.com/liip-forks/django-db-email-backend#egg=kn-django-db-email-backend
    # via -r /code/requirements/base.in
asgiref==3.7.2
    # via
    #   django
    #   django-countries
asttokens==2.4.0
    # via stack-data
backcall==0.2.0
    # via ipython
bcrypt==4.0.1
    # via paramiko
beautifulsoup4==4.12.2
    # via django-bootstrap3
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via
    #   dulwich
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via pip-tools
coverage[toml]==7.3.2
    # via pytest-cov
crispy-bootstrap3==2022.1
    # via -r /code/requirements/base.in
cryptography==41.0.4
    # via
    #   -r /code/requirements/base.in
    #   paramiko
    #   pyjwt
csscompressor==0.9.5
    # via -r /code/requirements/base.in
decorator==5.1.1
    # via
    #   fabric
    #   ipdb
    #   ipython
defusedxml==0.7.1
    # via
    #   odfpy
    #   python3-openid
deprecated==1.2.14
    # via fabric
diff-match-patch==20230430
    # via django-import-export
dj-database-url==2.1.0
    # via -r /code/requirements/base.in
django==4.2.6
    # via
    #   -r /code/requirements/base.in
    #   crispy-bootstrap3
    #   dj-database-url
    #   django-agpl
    #   django-allauth
    #   django-appconf
    #   django-autocomplete-light
    #   django-bootstrap3
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-ical
    #   django-import-export
    #   django-localflavor
    #   django-memoize
    #   django-parler
    #   django-phonenumber-field
    #   django-recurrence
    #   django-role-permissions
    #   django-taggit
    #   kn-django-db-email-backend
django-allauth==0.57.0
    # via -r /code/requirements/base.in
django-appconf==1.0.5
    # via django-compressor
django-autocomplete-light==3.9.7
    # via -r /code/requirements/base.in
django-bootstrap3==23.4
    # via -r /code/requirements/base.in
django-bootstrap3-datetimepicker-2==2.8.3
    # via -r /code/requirements/base.in
django-compressor==4.4
    # via -r /code/requirements/base.in
django-countries==7.5.1
    # via -r /code/requirements/base.in
django-crispy-forms==2.0
    # via
    #   -r /code/requirements/base.in
    #   crispy-bootstrap3
django-debug-toolbar==4.2.0
    # via -r requirements/dev.in
django-extensions==3.2.3
    # via -r requirements/dev.in
django-filter==23.3
    # via -r /code/requirements/base.in
django-ical==1.9.2
    # via -r /code/requirements/base.in
django-import-export==3.3.1
    # via -r /code/requirements/base.in
django-localflavor==4.0
    # via -r /code/requirements/base.in
django-memoize==2.3.1
    # via -r /code/requirements/base.in
django-parler==2.3
    # via -r /code/requirements/base.in
django-phonenumber-field==7.2.0
    # via -r /code/requirements/base.in
django-recurrence==1.11.1
    # via django-ical
django-role-permissions==3.2.0
    # via -r /code/requirements/base.in
django-simple-history==3.4.0
    # via -r /code/requirements/base.in
django-stronghold==0.4.0
    # via -r /code/requirements/base.in
django-taggit==4.0.0
    # via -r /code/requirements/base.in
django-tinymce==3.6.1
    # via -r /code/requirements/base.in
dulwich==0.19.15
    # via -r requirements/dev.in
et-xmlfile==1.1.0
    # via openpyxl
executing==2.0.0
    # via stack-data
fabric==3.2.2
    # via -r requirements/dev.in
factory-boy==3.3.0
    # via -r /code/requirements/test.in
faker==19.8.0
    # via factory-boy
icalendar==5.0.10
    # via django-ical
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
invoke==2.2.0
    # via fabric
ipdb==0.13.13
    # via -r /code/requirements/test.in
ipython==8.16.1
    # via ipdb
jedi==0.19.1
    # via ipython
libsass==0.22.0
    # via -r /code/requirements/base.in
markuppy==1.14
    # via tablib
markupsafe==2.1.3
    # via werkzeug
matplotlib-inline==0.1.6
    # via ipython
oauthlib==3.2.2
    # via requests-oauthlib
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
packaging==23.2
    # via
    #   build
    #   pytest
paramiko==3.3.1
    # via fabric
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
phonenumbers==8.13.22
    # via -r /code/requirements/base.in
pickleshare==0.7.5
    # via ipython
pillow==10.0.1
    # via -r /code/requirements/base.in
pip-tools==7.3.0
    # via -r requirements/dev.in
pluggy==1.3.0
    # via pytest
prompt-toolkit==3.0.39
    # via ipython
psycopg2-binary==2.9.9
    # via -r /code/requirements/base.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.16.1
    # via ipython
pyjwt[crypto]==2.8.0
    # via django-allauth
pynacl==1.5.0
    # via paramiko
pyproject-hooks==1.0.0
    # via build
pytest==7.4.2
    # via
    #   -r /code/requirements/test.in
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r /code/requirements/test.in
pytest-django==4.5.2
    # via -r /code/requirements/test.in
python-dateutil==2.8.2
    # via
    #   django-recurrence
    #   faker
    #   icalendar
python-stdnum==1.19
    # via django-localflavor
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3.post1
    # via
    #   -r /code/requirements/base.in
    #   icalendar
    #   kn-django-db-email-backend
pyyaml==6.0.1
    # via
    #   -r /code/requirements/base.in
    #   tablib
rcssmin==1.1.1
    # via django-compressor
requests==2.31.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
rjsmin==1.2.1
    # via django-compressor
ruff==0.12.0
    # via -r /code/requirements/test.in
sentry-sdk==1.31.0
    # via -r /code/requirements/base.in
six==1.16.0
    # via
    #   -r /code/requirements/base.in
    #   asttokens
    #   django-autocomplete-light
    #   python-dateutil
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via
    #   -r /code/requirements/base.in
    #   django-import-export
traitlets==5.11.2
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.8.0
    # via
    #   dj-database-url
    #   django-countries
urllib3==2.0.6
    # via
    #   dulwich
    #   requests
    #   sentry-sdk
wcwidth==0.2.8
    # via prompt-toolkit
werkzeug==3.0.0
    # via -r requirements/dev.in
wheel==0.41.2
    # via pip-tools
wrapt==1.15.0
    # via deprecated
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
